LayoutTests/svg/dynamic-updates/SVGFEConvolveMatrixElement-svgdom-targetY-prop-expected.txt

CONSOLE MESSAGE: line 41: TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.targetY.baseVal = 0')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'targetY' property of the SVGFEConvolveMatrixElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL convolveMatrixElement.targetY.baseVal should be 2. Threw exception TypeError: 'undefined' is not an object (evaluating 'convolveMatrixElement.targetY.baseVal')
